// src/app/admin/employees/shiftsNeedHelp/types.ts
export interface NeedHelpShift {
  id: string;
  departmentId: string;
  departmentName: string;
  employee: string;
  employeeId: string;
  employeeList: {
    id: string;
    userId: string;
    employee: string;
    phone: string;
    location: string;
    note: string;
    shiftDateTime: string;
    status: boolean;
  }[];
  roleId: string;
  roleName: string;
  locationId: string;
  locationName: string;
  proposer: string;
  proposerId: string;
  proposee: string[];
  status: NeedHelpShiftStatusVo; // "PENDING" | "APPROVED" | "REJECTED"
  partialShift: string; // "Yes"|"No"
  shiftId?: string;
  shiftDate: string;
  shiftEndTime: string;
  shiftStartTime: string;
  coverageStartTime?: string;
  coverageEndTime?: string;
  claimedBy: string;
  claimUserId?: string;
  phone: string;
  nationalPhone?: string;
  note: string;
  reason: string;
  comment: string;
  scheduleWeek: string;
  scheduleStartDate?: string;
  scheduleEndDate?: string;
  tags?: number[] | null;
}

export type NeedHelpShiftStatusVo =
  | "NEW"
  | "PENDING_APPROVAL"
  | "PENDING"
  | "APPROVED"
  | "REJECTED"
  | "APPROVE"
  | "REJECT"
  | "REQUESTED";